Some corrections in DPP implement (part3-Allocate Protection Calls )
Some adjusts were done in DPP functions and the Beta parameter was implemented.

+++ include/Data/Parameters.h
@@ -295,7 +295,7 @@ private:
      * @brief Squeezing index (beta) used in the simulation with Protection 
      * option enabled.
      */
-    unsigned int beta;
+    double beta;
     /**
      * @brief Calculate the load step.
      */
